renamed movie modal cards
