+++ src/index.html
@@ -22,6 +22,6 @@
     <!--footer-->
     <include src="./partials/footer.html"></include>
     <!-- modal-->
-    <include src="./partials/modal.html"></include>
+    <include src="./partials/modal_movie_cards.html"></include>
   </body>
 </html>
