+++ src/js-partials/open-close-modal_movie_card.js
