+++ src/sass/_modal_movie_cards.scss
