+++ src/sass/index.scss
@@ -5,5 +5,5 @@
 @import './loader';
 @import './gallery';
 @import './scrolltotop';
-@import "./modal";
+@import "./modal_movie_cards";
